ml_project/requirements.txt
antlr4-python3-runtime==4.9.3
dataclasses==0.6
hydra-core==1.2.0
joblib==1.2.0
mlopshw01==0.1.0
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
src==0.1.0
threadpoolctl==3.1.0
typing==3.7.4.3
